% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptions.R
\name{atlas_anchor}
\alias{atlas_anchor}
\title{Create stable HTML anchors}
\usage{
atlas_anchor(...)
}
\arguments{
\item{...}{Values to combine.}
}
\value{
A lowercase HTML-safe identifier.
}
\description{
Create stable HTML anchors
}
