% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema-model.R
\name{dependency_tree}
\alias{dependency_tree}
\title{Build a dependency tree}
\usage{
dependency_tree(model, entry_point, max_depth = Inf)
}
\arguments{
\item{model}{A \code{dbschema_model}.}

\item{entry_point}{A \code{schema.table} or \code{schema.table.column} identifier.}

\item{max_depth}{Maximum traversal depth.}
}
\value{
A \code{data.table} describing the traversal.
}
\description{
Traverses both dependencies and reverse usages without revisiting tables.
}
