% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema-model.R
\name{model_accessors}
\alias{model_accessors}
\alias{model_schemas}
\alias{model_tables}
\alias{model_columns}
\alias{model_dependencies}
\alias{model_usages}
\title{Access schema-model tables}
\usage{
model_schemas(model)

model_tables(model)

model_columns(model)

model_dependencies(model)

model_usages(model)
}
\arguments{
\item{model}{A \code{dbschema_model}.}
}
\value{
A copy of the requested \code{data.table}.
}
\description{
Access schema-model tables
}
