% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptions.R
\name{parse_description}
\alias{parse_description}
\title{Parse language-tagged descriptions}
\usage{
parse_description(x)
}
\arguments{
\item{x}{One character value containing blocks such as \verb{[EN] Text}.}
}
\value{
A named list of trimmed descriptions. Untagged text is returned as \code{EN}.
}
\description{
Parse language-tagged descriptions
}
