% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-io.R
\name{read_metadata}
\alias{read_metadata}
\title{Load a metadata snapshot}
\usage{
read_metadata(path)
}
\arguments{
\item{path}{Directory created by \code{\link[=write_metadata]{write_metadata()}}.}
}
\value{
A \code{dbschema_metadata} object.
}
\description{
Load a metadata snapshot
}
