% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{dependency_graph}
\alias{dependency_graph}
\title{Create an interactive dependency graph}
\usage{
dependency_graph(model, config = atlas_report_config())
}
\arguments{
\item{model}{A \code{dbschema_model}.}

\item{config}{An \code{\link[=atlas_report_config]{atlas_report_config()}} object.}
}
\value{
A \code{visNetwork} htmlwidget.
}
\description{
Create an interactive dependency graph
}
